core/requirements.txt
Flask
requests
python-dotenv
openai